+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All changes to `doctl` will be documented in this file.
 
 ## [1.13.1] - UNRELEASED
 
+- #408 k8s: Fix case where kube.Get returns a nil cluster
 - #398 Link to docs to create a Github token - @bouk
 - #392 Simplify newline trimming in retrieveUserInput - @timoreimann
 - #401 k8s: Fetch credentials after cluster is provisioned - @bouk